# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: RedPoint Monitor API
  version: 0.2.0
  description: |
    约定（设计 19.0.2）：JSON，字段 snake_case；时间为 Unix 秒；字节、字节/秒、百分比（0～100）。
    错误统一为 `{"error": {"code", "message", "request_id", "details"}}`，错误码见 ErrorCode（设计 43.4）。
    每个响应都带 `X-Request-ID`。
servers:
  - url: /api/v1
security:
  - session: []
...
paths:
...
  /api-keys:
    get:
      tags: [auth]
      summary: 只读 API Key 列表（设计 45.2）；不含完整 Key
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: object
                required: [next_cursor, items]
                properties:
                  next_cursor: { type: string, description: 空表示没有更多 }
                  items: { type: array, items: { $ref: '#/components/schemas/APIKey' } }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [auth]
      summary: 创建只读 API Key（需在 10 分钟内重新验证过密码）；完整 Key 只在响应中出现这一次
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/APIKeyInput' }
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                type: object
                required: [key, api_key]
                properties:
                  key: { type: string, description: 'api_…，只显示这一次' }
                  api_key: { $ref: '#/components/schemas/APIKey' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api-keys/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    delete:
      tags: [auth]
      summary: 吊销 API Key（立即生效，记录保留）
      responses:
        '204': { description: 已吊销 }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /cloud-instances/{id}/server:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    put:
      tags: [cloud]
      summary: 关联或取消关联节点（设计 44.5）；记入审计
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [server_id], properties: { server_id: { type: [integer, 'null'], description: null 表示取消关联 } } }
      responses:
        '200':
          description: 关联后的实例
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloudInstance' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /servers:
    get:
      security: [{ session: [] }, { apiKey: [] }, { deviceToken: [] }]
      tags: [servers]
      summary: 节点列表（实时状态取自内存）
      description: >-
        一次返回全部节点（总览的统计与筛选需要全量），next_cursor 恒为空；每核使用率、监听端口与扩展指标只在详情中返回（设计 19.0.2）。
      responses:
        '200':
          description: 节点列表
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ServerView' } }
                  next_cursor: { type: string, description: 空表示没有更多 }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [servers]
      summary: 新建节点（状态：待安装），同时生成注册码（设计 19.11、27.2）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateServerRequest' }
      responses:
        '201':
          description: 已创建。完整注册码只在此返回一次
          headers:
            Location: { schema: { type: string }, description: 新节点的路径 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollCodeView' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /servers/{id}:
    parameters: [{ $ref: '#/components/parameters/ServerID' }]
    get:
      security: [{ session: [] }, { apiKey: [] }, { deviceToken: [] }]
      parameters:
        - { name: live, in: query, schema: { type: integer, enum: [1] }, description: '按需实时模式（设计 46.2）：Web 详情页可见时带 live=1，节点在随后 30 秒内以 2 秒采样；API Key 带此参数无效' }
      tags: [servers]
      summary: 单个节点（格式同列表中的一项）
      responses:
        '200':
          description: 节点
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerView' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [servers]
      summary: 修改节点信息（设计 19.5）
      description: 整体替换可编辑信息，字段同新建（enroll_ttl 忽略）；未提供的可选字段视为清空。不改变注册状态。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateServerRequest' }
      responses:
        '200':
          description: 修改后的节点
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerView' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
    delete:
      tags: [servers]
      summary: 删除节点及其全部历史数据，不可恢复
      description: 该节点的 Agent Token 随之删除。敏感操作：需在 10 分钟内调用过 /auth/reauth，否则 403 reauth_required（设计 17.4）。
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: vpsmon_session
      description: |
        Web 管理员会话 Cookie（设计 8.2、17.4）。修改类请求（非 GET / HEAD）还必须带 X-CSRF-Token 头，
        值来自登录或 /auth/me 响应中的 csrf_token，否则 403。
    apiKey:
      type: http
      scheme: bearer
      description: |
        只读 API Key（api_…，设计 45.2）：Web“账号 → API Key”中创建，只显示一次。只能访问标注了 apiKey 的读取类接口；
        可限定节点范围（范围外的节点按不存在处理，返回 404）；每个 Key 每分钟 120 次。只接受 Authorization 头。
    deviceToken:
      type: http
      scheme: bearer
      description: |
        App 设备 Access Token（dev_…，设计 12.5）：AK 配对获得，30 分钟有效，用 Refresh Token（rt_…）轮换。
        只能访问标注了 deviceToken 的接口：读取授权范围内的节点、指标、流量、告警，以及（AK 允许时）静音与维护；
        范围外的节点按不存在处理（404）。过期返回 401 token_expired（应刷新），设备被吊销返回 401 token_revoked（应重新配对）。
        每台设备每分钟 120 次。
    agentToken:
      type: http
      scheme: bearer
      description: Agent Token（agt_…），只能为 Token 绑定的节点上报（设计 23.2）
  parameters:
    ServerID:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  schemas:
...
    EnrollTTL:
      type: string
      enum: ['1h', '24h', '7d']
      default: 24h
      description: 注册码有效期（设计 27.2）
...
    CreateServerRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        expected_hostname: { type: string, description: 注册时核对 }
        expected_ipv4: { type: string, format: ipv4 }
        expected_ipv6: { type: string, format: ipv6 }
        group: { type: string, maxLength: 32 }
        note: { type: string, maxLength: 500 }
        provider: { type: string, maxLength: 64 }
        plan: { type: string, maxLength: 64 }
        region: { type: string, maxLength: 64 }
        country: { type: string, pattern: '^[A-Za-z]{2}$', description: 'ISO 3166-1 两位代码（大小写不敏感，保存为大写）；用于显示国旗' }
        bandwidth_mbps: { type: integer, minimum: 0, maximum: 1000000, description: 服务商标称带宽（端口速率），Mbps；0 或省略表示未填（设计 27.2） }
        report_interval_s: { type: integer, enum: [0, 5, 10, 15, 30, 60], description: 采样间隔秒数；0 或省略表示默认 10 秒（设计 4.2、6.1） }
        traffic_timezone: { type: string, example: America/Los_Angeles, description: '计费时区（IANA 名称）；空或省略表示面板时区。只影响之后写入的流量（设计 5.4）' }
        traffic_limit_gb: { type: number, minimum: 0, description: 按 traffic_unit 口径的 GB / GiB（设计 5.8），0 或省略表示不限 }
        traffic_unit: { type: string, enum: [decimal, binary], default: decimal, description: 'decimal：1 GB = 10^9 字节；binary：1 GiB = 2^30 字节（设计 5.8）' }
        traffic_factor: { type: number, minimum: 0.5, maximum: 2, default: 1, description: 统计系数，修正固定比例偏差（设计 5.7） }
        traffic_reset_day: { type: integer, minimum: 1, maximum: 31, default: 1 }
        traffic_count_mode: { type: string, enum: [sum, rx, tx, max], default: sum, description: 设计 1.2.4 }
        price: { type: number, minimum: 0, description: 续费价格，保留两位小数 }
        currency: { type: string, pattern: '^[A-Za-z]{3}$', description: ISO 4217；填写价格时必填 }
        billing_period:
          type: string
          enum: ['', monthly, quarterly, semiannually, annually, biennially, triennially, one_time]
        expire_date: { type: string, format: date, description: YYYY-MM-DD }
        enroll_ttl: { $ref: '#/components/schemas/EnrollTTL' }
        verify_mode: { type: string, enum: [warn, strict], default: warn, description: 核对严格程度（设计 27.6.3） }
...
    EnrollCodeView:
      type: object
      required: [server_id, server_name, enroll_state, enroll_code_hint, enroll_status, enroll_expires_at, install]
      properties:
        server_id: { type: integer }
        server_name: { type: string }
        enroll_state: { type: string, enum: [pending, enrolled] }
        enroll_code: { type: string, description: 完整注册码，只在新建与重新生成时返回 }
        enroll_code_hint: { type: string, example: 'ENR-7KQ2-****' }
        enroll_status: { type: string, enum: [ACTIVE, USED, REVOKED, EXPIRED, NONE] }
        enroll_expires_at: { type: integer, description: Unix 秒 }
        install: { $ref: '#/components/schemas/InstallCommand' }
    ServerView:
      type: object
      required: [id, name, status, enroll_state, expected_hostname, expected_ipv4, expected_ipv6, verify_mode, enrolled_at, created_at, hostname, ipv4, ipv6, group, note, provider, plan, region, country, bandwidth_mbps, report_interval_s, price_cents, currency, billing_period, expire_date, traffic_limit_bytes, traffic_reset_day, traffic_timezone, traffic_count_mode, traffic_unit, traffic_factor, last_seen_at, alerts, traffic]
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string, enum: [online, unknown, offline, pending], description: '≤30 秒在线，≤120 秒未知（设计 22）；pending 为待安装（设计 27.7）' }
        last_seen_at: { type: integer }
        enroll_state: { type: string, enum: [pending, enrolled] }
        expected_hostname: { type: string }
        expected_ipv4: { type: string }
        expected_ipv6: { type: string }
        verify_mode: { type: string, enum: [warn, strict] }
        hostname: { type: string }
        ipv4: { type: string }
        ipv6: { type: string }
        enrolled_at: { type: integer }
        group: { type: string }
        note: { type: string }
        provider: { type: string }
        plan: { type: string }
        region: { type: string }
        country: { type: string, description: ISO 3166-1 两位代码（大写），空表示未填 }
        bandwidth_mbps: { type: integer, description: 标称带宽 Mbps，0 表示未填 }
        report_interval_s: { type: integer, description: 采样间隔秒数，0 表示默认 10 秒 }
        traffic_timezone: { type: string, description: 计费时区（IANA），空表示面板时区 }
        price_cents: { type: integer, description: 续费价格 × 100 }
        currency: { type: string }
        billing_period: { type: string }
        expire_date: { type: string }
        traffic_limit_bytes: { type: integer }
        traffic_reset_day: { type: integer }
        traffic_count_mode: { type: string, enum: [sum, rx, tx, max] }
        created_at: { type: integer }
        latest: { $ref: '#/components/schemas/Report' }
        traffic_unit: { type: string, enum: [decimal, binary] }
        traffic_factor: { type: number }
        traffic: { $ref: '#/components/schemas/TrafficView' }
        alerts:
          type: array
          description: 活动告警（firing），严重在前；同一类型只保留最严重的一条；节点离线时只有离线告警（设计 16.4）
          items:
            type: object
            required: [event_id, rule_key, type, severity, message, value, fired_at, silenced]
            properties:
              event_id: { type: integer }
              rule_key: { type: string }
              type: { type: string }
              severity: { type: string, enum: [info, warning, critical] }
              message: { type: string }
              value: { type: number, description: 当前值：百分比、离线秒数或负载倍数 }
              fired_at: { type: integer }
              silenced: { type: boolean, description: 已静音：照常记录，不通知，不计入需要关注 }
        maintenance: { $ref: '#/components/schemas/Silence' }
        muted: { $ref: '#/components/schemas/Silence' }
...
    TrafficView:
      type: object
      description: 本计费周期流量（设计 5.7、5.8、32）。字节数均为整数，GB 换算由客户端按 unit 完成。
      required: [cycle_start, cycle_end, rx, tx, measured, adjustment, used, limit, unit, factor]
      properties:
        cycle_start: { type: string, format: date }
        cycle_end: { type: string, format: date, description: 下一周期开始日（不含） }
        rx: { type: integer }
        tx: { type: integer }
        measured: { type: integer, description: 统计值：按计费模式取值 × 系数，不含校准 }
        adjustment: { type: integer, description: 本周期最近一次校准在当前系数与计费模式下的偏差，可为负（设计 5.7） }
        calibrated_at: { type: integer, description: 最近一次校准时间，Unix 秒；未校准时省略 }
        used: { type: integer, description: 展示值 = 统计值 + 校准偏差 }
        limit: { type: integer, description: 字节，0 表示不限 }
        unit: { type: string, enum: [decimal, binary] }
        factor: { type: number }
        forecast:
          type: object
          description: 周期结束时的预计用量；周期开始不足 3 天时省略（设计 32）
          required: [daily, total, over]
          properties:
            daily: { type: integer, description: 采用的日均 }
            total: { type: integer }
            over: { type: boolean, description: 预计超过额度 }
        factor_suggestion: { type: number, description: 多次校准显示稳定的比例偏差时建议设置的统计系数；没有建议时省略（设计 5.7） }
    Report:
      type: object
      description: Agent 上报，定义见 internal/protocol（设计 6.2）；只做向后兼容的新增
      required: [timestamp, agent_version, system, cpu, memory, swap, disk, network]
      properties:
        timestamp: { type: integer }
        agent_version: { type: string }
        final: { type: boolean }
        sent_at: { type: integer, description: 可选：发送时刻的 Agent 时钟（Unix 秒），面板据此计算时钟偏差（设计 43.5） }
        system:
          type: object
          required: [hostname, os, os_version, kernel, arch, uptime]
          properties:
            hostname: { type: string }
            os: { type: string }
            os_version: { type: string }
            kernel: { type: string }
            arch: { type: string }
            uptime: { type: integer, description: 秒 }
            boot_id: { type: string }
            cpu_model: { type: string, description: 可选（设计 4.4） }
            counter_bits: { type: integer, enum: [32, 64], description: 可选：内核网卡计数器位数；32 位时面板按回绕补算流量（设计 5.5） }
            remote_upgrade: { type: boolean, description: 可选：本机是否启用了远程升级（设计 29.13）；省略表示旧版 Agent、未知 }
        cpu:
          type: object
          required: [usage, cores, load1]
          properties:
            usage: { type: number }
            cores: { type: integer }
            load1: { type: number }
            load5: { type: number }
            load15: { type: number }
            breakdown:
              type: object
              description: 可选：两次采样之间各类时间占比 0～100（设计 4.4）
              required: [user, nice, system, iowait, irq, softirq, steal, idle]
              properties:
                user: { type: number }
                nice: { type: number }
                system: { type: number }
                iowait: { type: number }
                irq: { type: number }
                softirq: { type: number }
                steal: { type: number }
                idle: { type: number }
            per_core: { type: array, items: { type: number }, description: 可选；GET /servers 列表中省略 }
            temp_c: { type: number, description: 可选，CPU 温度 ℃ }
        memory:
          type: object
          required: [total, used, usage]
          properties:
            total: { type: integer }
            used: { type: integer, description: MemTotal − MemAvailable }
            available: { type: integer }
            usage: { type: number }
            free: { type: integer, description: 可选 }
            buffers: { type: integer, description: 可选 }
            cached: { type: integer, description: 可选：Cached + SReclaimable }
        swap:
          type: object
          required: [total, used]
          properties:
            total: { type: integer }
            used: { type: integer }
        processes:
          type: object
          description: 可选（设计 4.9）
          required: [total, running]
          properties:
            total: { type: integer }
            running: { type: integer }
        conns:
          type: object
          description: 可选（设计 4.9）：sockstat 的 inuse，IPv4 + IPv6
          required: [tcp, udp, time_wait]
          properties:
            tcp: { type: integer }
            udp: { type: integer }
            time_wait: { type: integer }
        ports:
          type: array
          maxItems: 64
          description: 可选（设计 4.9.1）：本机监听端口，按协议与端口合并；不含连接明细与进程信息。节点列表接口省略此字段
          items:
            type: object
            required: [proto, port, addrs]
            properties:
              proto: { type: string, enum: [tcp, udp] }
              port: { type: integer }
              addrs: { type: array, items: { type: string }, description: '监听地址，如 0.0.0.0、::、127.0.0.1' }
        disk:
          type: array
          description: 每个挂载点（设计 4.6）；只采集本地块设备文件系统，同一设备只报一次
          items:
            type: object
            required: [mount, total, used, usage]
            properties:
              mount: { type: string }
              total: { type: integer }
              used: { type: integer }
              usage: { type: number, description: 与 df 的 Use% 一致 }
              fstype: { type: string }
              device: { type: string }
              available: { type: integer }
        network:
          type: array
          description: 每块网卡；rx/tx_bytes 为内核累计字节数，rx/tx_speed 为 Agent 计算的字节/秒
          items:
            type: object
            required: [interface, rx_bytes, tx_bytes, rx_speed, tx_speed]
            properties:
              interface: { type: string }
              ifindex: { type: integer }
              rx_bytes: { type: integer }
              tx_bytes: { type: integer }
              rx_speed: { type: integer }
              tx_speed: { type: integer }
              rx_pps: { type: number, description: 可选：包速率（设计 4.10） }
              tx_pps: { type: number }
              rx_errors: { type: integer, description: 可选：开机以来的错误、丢包累计 }
              tx_errors: { type: integer }
              rx_dropped: { type: integer }
              tx_dropped: { type: integer }
        disk_io:
          type: array
          description: 可选，整块磁盘的 IO（设计 4.7）
          items:
            type: object
            required: [device, read_speed, write_speed]
            properties:
              device: { type: string }
              read_bytes: { type: integer }
              write_bytes: { type: integer }
              read_ops: { type: integer }
              write_ops: { type: integer }
              io_time_ms: { type: integer }
              read_speed: { type: integer }
              write_speed: { type: integer }
              read_iops: { type: number, description: 可选 }
              write_iops: { type: number, description: 可选 }
              await_ms: { type: number, description: 可选：每次 IO 平均耗时 }
              util: { type: number, description: 可选：设备忙碌占比 0～100 }
              in_flight: { type: integer, description: 可选：采样时正在处理的 IO 请求数（设计 4.10） }
        collect_errors:
          type: array
          description: 可选：本轮失败的采集项及原因，对应字段留空，其余照常（设计 43.5）
          items:
            type: object
            required: [item, message]
            properties:
              item: { type: string, enum: [system, cpu, memory, disk, disk_io, network, processes, conns, ports, extra] }
              message: { type: string }
        extra:
          type: object
          description: >-
            可选：扩展指标（设计 4.10），字段定义见 internal/protocol.Extra。面板保存在实时状态中随接口返回，暂不入库也不展示。
            读不到或内核不提供的项省略；速率为与上一次采样之间的平均值。
          properties:
            activity: { type: object, description: /proc/stat：ctx_switches、interrupts、forks（每秒）、procs_blocked }
            vm: { type: object, description: /proc/vmstat：major_faults（每秒）、swap_in / swap_out（字节/秒）、oom_kills（开机以来，内核 4.13+） }
            pressure: { type: object, description: PSI（内核 4.20+）：cpu_some_avg10/60、memory_some_avg10、memory_full_avg10、io_some_avg10、io_full_avg10 }
            memory_more: { type: object, description: /proc/meminfo：shmem、slab_unreclaim、dirty、writeback、committed、commit_limit（字节） }
            net: { type: object, description: /proc/net/snmp：tcp_established、tcp_retrans_rate（%）、tcp_active_opens / tcp_passive_opens（每秒）、错误累计 }
            file_handles: { type: object, description: /proc/sys/fs/file-nr：allocated、max }
            conntrack: { type: object, description: nf_conntrack：count、max；未加载时省略 }
            clock: { type: object, description: adjtimex：synced、max_error_us }
            env:
              type: object
              description: 运行环境（设计 46.3）
              properties:
                virt: { type: string, description: 'kvm / xen / vmware / hyperv / virtualbox / openvz / lxc / docker / podman / wsl / vm / none；无法判断时省略' }
                dmi_vendor: { type: string, description: /sys/class/dmi/id/sys_vendor，如 QEMU、Alibaba Cloud }
                dmi_product: { type: string }
...
    CloudInstance:
      type: object
      required: [id, account_id, account_name, provider, instance_id, name, region, kind, state, public_ipv4, public_ipv6, plan,
                 expire_at, renew_price_cents, traffic_limit_bytes, traffic_used_bytes, traffic_period_start, server_id, updated_at,
                 suggested_server_id]
      properties:
        id: { type: integer }
        account_id: { type: integer }
        account_name: { type: string }
        provider: { type: string }
        instance_id: { type: string }
        name: { type: string }
        region: { type: string }
        kind: { type: string, description: ec2 / lightsail / ecs / swas（阿里云轻量）/ cvm / lighthouse（腾讯云轻量）/ oci / oci_egress（Oracle 租户每月出站流量） }
        state: { type: string, description: 云厂商原样，如 running / stopped }
        public_ipv4: { type: string }
        public_ipv6: { type: string }
        plan: { type: string, description: 规格（EC2 实例类型、Lightsail 套餐） }
        expire_at: { type: integer, description: 到期时间，Unix 秒；按需付费为 0 }
        renew_price_cents: { type: integer }
        traffic_limit_bytes: { type: integer, description: 流量包额度，字节；0 表示没有流量包 }
        traffic_used_bytes: { type: integer, description: 本周期已用（云厂商口径，有数小时延迟） }
        traffic_period_start: { type: string, description: YYYY-MM-DD }
        server_id: { type: [integer, 'null'], description: 关联的节点（设计 44.5） }
        suggested_server_id: { type: [integer, 'null'], description: 未关联时，公网 IP 与之唯一匹配的节点（建议关联，需用户确认） }
        updated_at: { type: integer }
    APIKeyInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        scope_type: { type: string, enum: [all, group, servers], default: all }
        group: { type: string, description: scope_type 为 group 时的分组名 }
        server_ids: { type: array, items: { type: integer }, maxItems: 1000, description: scope_type 为 servers 时的节点 }
        expires_in_days: { type: integer, minimum: 0, maximum: 3650, description: 0 表示不过期 }
    APIKey:
      type: object
      required: [id, name, hint, scope_type, scope_value, expires_at, created_by, created_at, last_used_at, revoked_at]
      properties:
        id: { type: integer }
        name: { type: string }
        hint: { type: string, description: Key 的末 4 位 }
        scope_type: { type: string, enum: [all, group, servers] }
        scope_value: { type: string, description: 分组名，或逗号分隔的节点 ID }
        expires_at: { type: integer, description: 过期时间，Unix 秒；0 表示不过期 }
        created_by: { type: string }
        created_at: { type: integer }
        last_used_at: { type: integer, description: 最近使用时间（每分钟最多更新一次）；0 表示从未使用 }
        revoked_at: { type: integer, description: 吊销时间；0 表示有效 }
...